Created the single model for the surveys
I have worked out that we can use a single database for all of the surveys by using a few
identifiers.  We can use a survey type & subtype to indicate if it is a Pollinator (type) survey and
a Bee, Ant, Bird, etc. (subtype) survey.

We also associate a question set (integer value) that, in combination with the survey type, is used
to determine which questions were asked during the survey. This allows us to dynamically add more
questions, by simply adding more question sets. This also allows database modifications to be made
easier since, most likely, the only structural change that will have to be made is adding more
columns to support a greater quantity of questions per survey.

By doing all of this, it allows us to generalize the surveys that are submitted while simplifying
our database structure